refactor(workflow): split workflow.go into 2 focused files
